requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
aiohttp>=3.9.1
asyncio>=3.4.3
typing-extensions>=4.8.0
greenlet